STNDP-29 Integrate Settings > Timezone
